Refactor Plugins
